Use active binding to retrieve internal tiledb object
Slight re-working of the initial strategy: `object` is now a public
active binding. This approach makes it possible to check whether
private$tiledb_object has been populated yet and, if not, to call the
private$initialize_object() method. This is necessary for
TileDBArray objects, since an array is created after initializing the
class.