Remove special handling for adding scaled.data to coo
The count, data, and scaled.data slots of a Seurat assay are no longer
stored in different attributes of a single array. Instead, they are stored
in different arrays within the X group. Accordingly, we no longer need
this code that merged two COO dataframes with unequal coordinates.
I'm living in place the code for consolidating multiple "layerable"
matrices into a single COO dataframe, which might be might be
leveraged in the future.